separate axial, coronal, and sagittal views into different classes
